requirements.txt
# AgentMark ALFWorld 实验环境依赖
# Python >= 3.9
# Install dependencies: pip install -r requirements.txt

# OpenAI API
openai>=1.0.0
tqdm>=4.66.0

# PyTorch (CPU版本，ALFWorld文本环境不需要GPU)
torch>=2.0.0

# 数据处理和科学计算
numpy>=1.24.0

# 可视化
matplotlib>=3.7.0

# 其他工具
pyyaml>=6.0
pandas>=1.5.0
scipy>=1.10.0
requests>=2.31.0
tenacity>=8.2.0
backoff>=2.2.0

# 深度学习与大模型评估
transformers>=4.30.0
accelerate>=0.20.0
vllm>=0.2.0

# 注意：ALFWorld 在 Docker 中单独安装，本地环境请运行：
# pip install alfworld
